Homework for the lesson "Library Room" - development

+++ app/src/main/java/ru/netology/nmedia/repository/PostRepository.kt
@@ -1,13 +1,13 @@
 package ru.netology.nmedia.repository
 
 import androidx.lifecycle.LiveData
-import ru.netology.nmedia.dto.Post
+import ru.netology.nmedia.model.dto.Post
 
 interface PostRepository {
     fun getAll(): LiveData<List<Post>>
-    fun likeById(id: Int)
-    fun shareById(id: Int)
-    fun viewingById(id: Int)
+    fun likeById(id: Long)
+    fun shareById(id: Long)
+    fun viewingById(id: Long)
     fun savePost(post: Post)
-    fun removeById(id: Int)
+    fun removeById(id: Long)
 }
